section 5 (partial): extract archive edit orchestration
src/model/archive_edit.rs (243 lines) owns ArchiveEditContext,
ArchiveEditProgress, start_extract, start_compress, and the shared
compute_file_checksum. Each launcher spawns the worker thread and returns
an Arc<Mutex<ArchiveEditProgress>> for the GUI to poll.

browse_view.rs::start_archive_extract (~60 lines) and start_archive_compress
(~110 lines) collapsed to ~20 lines each — view-side state mutations stay
on the view, the model handles the threading and disk I/O. browse_view.rs
shrinks 4058 -> 3851 lines.
