+++ src/model/mod.rs
@@ -3,4 +3,5 @@
 //! The eventual goal is for every GUI tab to be a renderer over an object in
 //! this tree. See §5 / §10 of `docs/codecleanup.md`.
 
+pub mod archive_edit;
 pub mod edit_queue;
